requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
catboost==1.2.5
matplotlib==3.8.4
seaborn==0.13.2
flask==3.0.3
streamlit==1.35.0
shap==0.44.1
joblib==1.4.2
PyYAML==6.0.1
